bridging back and forth works

+++ web/packages/operations/src/transfer_weth_from_kusama_to_polkadot.ts
@@ -63,11 +63,10 @@ const transfer = async () => {
         },
     })
 
-    console.log("CONNECTING TO POLKADOT AH")
     const [polkadotAssetHub] = await Promise.all([
         context.assetHub(),
     ])
-    console.log("CONNECTING TO KUSAMA AH")
+
     const [kusamaAssetHub] = await Promise.all([
         context.kusamaAssetHub(),
     ])
@@ -129,16 +128,16 @@ const transfer = async () => {
         }
 
         // Step 5. Submit transaction and get receipt for tracking
-        //const response = await toKusama.signAndSend(
-        //    kusamaAssetHub,
-        //    transfer,
-        //    SUBSTRATE_ACCOUNT,
-        //    { withSignedTransaction: true }
-        //)
-        //if (!response) {
-        //    throw Error(`Transaction ${response} not included.`)
-            //}
-        //console.log("Success message", response.messageId)
+        const response = await toKusama.signAndSend(
+            kusamaAssetHub,
+            transfer,
+            SUBSTRATE_ACCOUNT,
+            { withSignedTransaction: true }
+        )
+        if (!response) {
+            throw Error(`Transaction ${response} not included.`)
+        }
+        console.log("Success message", response.messageId)
 
         await context.destroyContext()
     }
